fix: ssr error
